refactor frontend directory structure and adjust imports in file to use global app path instead of relative imports addresses #32
