+++ frontend/app/lib/components/hero.tsx
