+++ frontend/app/lib/components/index.ts
@@ -1,2 +1,3 @@
 export * from './client-side';
+export * from './hero';
 export * from './project-card';
