+++ frontend/app/lib/components/project-card.tsx
@@ -1,4 +1,4 @@
-import { Project, languages } from "../types"
+import { Project, languages } from "@/app/lib/types"
 import Link from "next/link";
 import LanguageBar from "./language-bar/language-bar";
 
@@ -21,20 +21,20 @@ const determineLanguagePercents = (githubPayload: {
   return percents;
 }
 
-export default function ProjectCard(props: Omit<Project, "filterableMetadata" | "account"> & {githubPayload: {[K in typeof languages[number]]?: number}}) {
+export function ProjectCard(props: Omit<Project, "filterableMetadata" | "account"> & {githubPayload: {[K in typeof languages[number]]?: number}}) {
   const languagePercents = determineLanguagePercents(props.githubPayload);
 
   return(
     <Link href={props.href}>
-    <div className="m-2 p-2 w-auto h-auto sm:min-w-[340px] sm:min-h-[300px] sm:max-w-[340px] border border-gray-300 rounded-md hover:shadow-lg flex flex-col justify-between">
+    <div className="flex flex-col justify-between border-gray-300 hover:shadow-lg m-2 p-2 border rounded-md w-auto sm:min-w-[340px] sm:max-w-[340px] h-auto sm:min-h-[300px]">
         
       <div id="title"
-        className="p-2 flex justify-center text-center font-bold">
+        className="flex justify-center p-2 font-bold text-center">
         <div>
         {props.displayTitle}
         </div>
       </div>
-      <div className="px-2 font-light text-center text-base justify-center flex">
+      <div className="flex justify-center px-2 font-light text-base text-center">
         <div>
         {props.description}
         </div>
